moving and clearing SHP2Array to com2AB folder
